Updated SuspectState
Updated the action that the agent/NPC takes when the player collides with an object.

The NPC will continue to go to the object until it is at the location (i.e. when NPC and object distance is 0). When it is, it will set soundDetected field to false, allowing NPC to switch state).