Added a TODO comment For Future Work
Added a TODO comment to remind myself of an issue that I had found.

When the agent goes to into suspect state, if nothing happens and the agent goes back to patrol state, the agent sticks to patrol state even if the player is right in front of the agent.