Fixed State Change Issue
Fixed the issue. Turns out it wasn't an issue with the FSM, but an issue with the player themselves. It looks like everytime the player moves and rotates, the player moves ever so slightly upwards. Meaning, if the player moves long enough, they will be too high up for the NPC to detect their presence (i.e. their hitbox is above the raycast).