Edited Player and BaseTest Prefabs
When the player moved, sometimes the player would zip through the walls (which we want to avoid).

To avoid this, I changed the players mass to 1 and the BaseTest CollisionDetection to be "Continuous Dynamic". With a smaller mass, the player will never zip through the object and with continuous dynamic, the game will continuously check for collision of objects (even fast objects).