Removed Player Rotation Mechanic
Removed rotation mechanic. Now, if the player clicks the left or right button, the player model will rotate 90 degrees left or right as long as the previous button pressed is not the same button as the button you are currently pressing.

This is to help the player avoid from turning continuously if they are holding down the left or right button.