Modified Gradient Change
Should've added a description for previous commit:

The color of the map now changes gradually and continuously.

Efficiency wise, the way it is working now is absolutely abysmal (overrides previous color and creates new material every single time).

This commit:
Changed the color increment so the change looks more gradual.