Added Upward Generation Beam Obstacle
Added an obstacle where 20 beams are generated randomly within the players vicinity. There is a 5 second indicator period where the "indicators" that are generated (circles on the ground) flicker between red and magenta to warn the player of an incoming obstacle. After the 5 seconds are done, beams will fire upwards from the indicators.

In addition, I changed the Boss bullets to be spheres instead of cylinders. Didn't want to spend too much time on trying to get the cylinders to rotate depending on X and Z axis.

Removed "Player Health Bar" folder. The player health system is stored in the "Player Script" folder instead.

I plan on working on a "sideways" obstacle as well, where the indicators are on the walls instead. That, or I could simply generate walls that slowly creep closer to the player. Regardless, there will be 1 other obstacle that the player will account for.

Lastly, I will implement a "duration system". Basically, it'll help inform the player how long they need to survive in the battle mode before it ends.